+++ server/src/serverconfig.js
@@ -160,6 +160,8 @@ if (serverconfig.allow_env_overrides) {
 	}
 }
 
+if (serverconfig.URL?.endsWith('/')) serverconfig.URL = serverconfig.URL.slice(0, -1)
+
 // always change selected configuration paths in a container
 if (process.env.PP_MODE?.startsWith('container')) {
 	// within the container, the Dockerfile uses pre-determined port and filepaths